backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.20.0
pydantic>=1.10.0
httpx>=0.23.0
aiohttp>=3.8.4
python-dotenv>=1.0.0
python-multipart>=0.0.6
starlette>=0.26.1
anyio>=3.6.2
asyncio>=3.4.3
typing-extensions>=4.5.0
sqlalchemy==2.0.20
alembic==1.12.0
pymysql==1.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic_settings==2.8.1